feat(#2): moved console to separate folder
* preparation for variable/command management
